Move table init upstream
