+++ packages/plugins/plugin-table/src/hooks/useTableModel.ts
@@ -4,10 +4,8 @@
 
 import { useEffect, useState } from 'react';
 
-import { create, S, TypedObject } from '@dxos/echo-schema';
-import { PublicKey } from '@dxos/react-client';
-import { type EchoReactiveObject, getSpace } from '@dxos/react-client/echo';
-import { createView, type ViewProjection } from '@dxos/schema';
+import { type EchoReactiveObject } from '@dxos/react-client/echo';
+import { type ViewProjection } from '@dxos/schema';
 
 import { TableModel, type TableModelProps } from '../model';
 import { type TableType } from '../types';
@@ -25,34 +23,6 @@ export const useTableModel = ({
   onDeleteColumn,
   onDeleteRow,
 }: UseTableModelParams): TableModel | undefined => {
-  const space = getSpace(table);
-
-  // TODO(burdon): Move up-stream.
-  useEffect(() => {
-    if (space && table && !table?.view) {
-      const schema = TypedObject({
-        typename: 'example.com/type/' + PublicKey.random().truncate(),
-        version: '0.1.0',
-      })({
-        name: S.optional(S.String),
-        description: S.optional(S.String),
-        quantity: S.optional(S.Number),
-      });
-
-      // Register schema.
-      const mutable = space.db.schemaRegistry.addSchema(schema);
-      table.view = createView({
-        typename: mutable.typename,
-        jsonSchema: mutable.jsonSchema,
-        properties: ['name', 'description', 'quantity'],
-      });
-
-      // Create first row.
-      space.db.add(create(mutable, {}));
-    }
-  }, [space, table]);
-
-  // Create model.
   const [model, setModel] = useState<TableModel>();
   useEffect(() => {
     if (!table || !projection) {
